refactor player.js
